refactor(ios): one PCM audio player and capturer
`WatchAudioPlayer` and `WatchAudioCapturer` differed from the phone's pair by
an audio-session policy and some comment wording — the watch configures no
session, because `WatchVoiceAudioSession` owns one for the whole call, and it
refuses a denied microphone before the tap begins, because the engine raises
there rather than failing.

Those three differences travel as one `PCMAudioSessionPolicy` value, so a
caller cannot pick a combination neither platform has, and the four files
become `PCMAudioPlayer` and `PCMAudioCapturer` in `LukeKit` behind an
iOS-and-watchOS guard. The two comments that record a bug — the simulator's
zero-channel format that raises an Objective-C exception inside `installTap`,
and the converter's repeat-call `.noDataNow` guard — are kept word for word.
